requirements.txt
GPlayCli
PyExecJS
PyGithub
aiohttp
colorthief
discord.py
event_bus
mongoengine
requests
wikipedia